fix: 컬럼/카드 컴포넌트 css수정, 컬럼 추가시 페이지 리로딩 없이 랜더링
