+++ src/components/atoms/column/column.module.scss
@@ -6,6 +6,7 @@
     width: 345px;
     border-right: 1px solid var(--gray30);
     padding: 20px;
+    border-bottom: none;
   }
 }
