+++ src/components/molecules/card/card.module.scss
@@ -15,6 +15,7 @@
   }
 }
 .cardImage {
+  border-radius: 6px;
   @media (min-width: 376px) and (max-width: 768px) {
     width: 90px;
     height: 53px;
