+++ src/components/molecules/cardSection/CardSection.tsx
@@ -1,37 +1,27 @@
 import styles from './cardSection.module.scss';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import CreateCardModal from '../modals/createCardModal/CreateCardModal';
 import EditColumnModal from '../modals/editColumnModal/EditColumnModal';
-import instance from '@/api/axios';
-
 import Column from '@/components/atoms/column/Column';
 
-interface CardSectionProps {
+interface Column {
+  id: number;
+  title?: string;
+  teamId: string;
   dashboardId: number;
+  createdAt?: string;
+  updatedAt?: string;
 }
 
-export default function CardSection({ dashboardId }: CardSectionProps) {
+interface CardSectionProps {
+  columns: [];
+  getColumns: () => void;
+}
+
+export default function CardSection({ columns, getColumns }: CardSectionProps) {
   const [isCreateCardModalOpen, setIsCreateCardModalOpen] = useState(false);
   const [isEditColumnModalOpen, setIsEditColumnModalOpen] = useState(false);
-  const [columns, setColumns] = useState<Column[]>([]);
   const [selectedColumn, setSelectedColumn] = useState<Column | null>(null);
-
-  async function getColumns() {
-    try {
-      const res = await instance.get<{ data: Column[] }>(
-        `/columns?dashboardId=${dashboardId}`,
-        {
-          headers: {
-            Accept: 'application/json',
-            Authorization: `Bearer ${localStorage.getItem('accessToken')}`,
-          },
-        },
-      );
-      setColumns(res.data.data);
-    } catch (error) {
-      console.error('Error fetching columns:', error);
-    }
-  }
 
   const handleAddCardButtonClick = (column: Column) => {
     setSelectedColumn(column);
@@ -48,10 +38,6 @@ export default function CardSection({ dashboardId }: CardSectionProps) {
     setIsCreateCardModalOpen(false);
     setIsEditColumnModalOpen(false);
   };
-
-  useEffect(() => {
-    getColumns();
-  }, [dashboardId]);
 
   return (
     <div className={styles['cardSection']}>
