+++ src/components/molecules/modals/createColumnModal/CreateColumnModal.tsx
@@ -7,6 +7,7 @@ import instance from '@/api/axios';
 
 interface ModalProps {
   onClose: () => void;
+  getColumns: () => void;
   dashboardId: number;
   setColumns: React.Dispatch<React.SetStateAction<any[]>>;
 }
@@ -15,6 +16,7 @@ export default function CreateColumnModal({
   onClose,
   dashboardId,
   setColumns,
+  getColumns,
 }: ModalProps) {
   const [columnName, setColumnName] = useState('');
 
@@ -36,7 +38,8 @@ export default function CreateColumnModal({
       });
       console.log('Column added successfully:', res.data);
       onClose();
-      window.location.reload();
+      // window.location.reload();
+      getColumns();
     } catch (error) {
       console.error('Error adding column:', error);
     }
